NS-99/core-module: Add core module. Move there all the core-related stuff
