+++ net-gateway/src/core/app_state.rs
