+++ net-gateway/src/core/client_data.rs
