+++ net-gateway/src/core/general_filters.rs
