+++ net-gateway/src/core/mod.rs
@@ -1,0 +1,3 @@
+pub mod app_state;
+pub mod client_data;
+pub mod general_filters;
