+++ net-gateway/src/lib.rs
@@ -1,5 +1,3 @@
 pub mod authorization;
-pub mod app_state;
+pub mod core;
 pub mod endpoints;
-pub mod client_data;
-pub mod general_filters;
